% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAgena.R
\name{create_batch_cases}
\alias{create_batch_cases}
\title{Create a json file of a BN model with observations}
\usage{
create_batch_cases(inputModel, inputData)
}
\arguments{
\item{inputModel}{an R model object}

\item{inputData}{CSV file of observations}
}
\value{
json file
}
\description{
This function takes an R Model objectand a CSV file of observations for cases
and creates a batch of datasets / cases for each row in the input data
and generates a .json file for the BN
}
