% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAgena.R
\name{create_sensitivity_config}
\alias{create_sensitivity_config}
\title{Create configuration object for sensitivity analysis}
\usage{
create_sensitivity_config(
  target,
  sensitivity_nodes,
  dataset = NULL,
  network = NULL,
  report_settings = NULL
)
}
\arguments{
\item{target}{target node in the network}

\item{sensitivity_nodes}{a set of sensitivity nodes in the network}

\item{dataset}{a dataSet in the model}

\item{network}{a network in the model}

\item{report_settings}{a list of report settings}
}
\value{
sensitivity config object
}
\description{
See agena.ai manual for sensitivity analyis configuration
}
