% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAgena.R
\docType{class}
\name{Dataset-class}
\alias{Dataset-class}
\alias{Dataset}
\title{Dataset}
\description{
Dataset
}
\section{Fields}{

\describe{
\item{\code{id}}{dataSet id}

\item{\code{observations}}{list ob observations.}

\item{\code{results}}{list of inference results.}
}}


