% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAgena.R
\name{from_cmpx}
\alias{from_cmpx}
\title{Load a Model from CMPX}
\usage{
from_cmpx(modelPath)
}
\arguments{
\item{modelPath}{Path to the input file}
}
\value{
An R model object
}
\description{
This function reads an input agena.ai model file with the .cmpx extension
to create an R model object. The model object includes the networks and
the nodes of the imported model.
}
