% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAgena.R
\name{local_api_activate_license}
\alias{local_api_activate_license}
\title{Local agena.ai developer license activation}
\usage{
local_api_activate_license(key)
}
\arguments{
\item{key}{agena.ai developer license key}
}
\description{
sends the agena.ai developer license for activation
}
