% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAgena.R
\name{local_api_batch_calculate}
\alias{local_api_batch_calculate}
\title{Local agena.ai API model calculation for batch cases}
\usage{
local_api_batch_calculate(model)
}
\arguments{
\item{model}{an R model object}
}
\value{
a json file for the model with results
}
\description{
Runs inference for every dataSet in the model
}
