% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAgena.R
\name{local_api_sensitivity}
\alias{local_api_sensitivity}
\title{Local agena.ai API sensitivity analysis}
\usage{
local_api_sensitivity(model, sens_config, output)
}
\arguments{
\item{model}{an R model object}

\item{sens_config}{a sensitivity config object}

\item{output}{file name for the output json}
}
\value{
a json file report of the sensitivity analysis results
}
\description{
Local agena.ai API sensitivity analysis
}
